Handle prefix search where search ends at leaf node without consuming.
Need to handle the case where the search traversal gets you to a leaf
node, but you haven't consumed the whole key. An example is the trivial
case of one string and an empty prefix -- only a leaf node will be
present, and the prefix doesn't consume the string.

To handle this case I had the search helper return the leaf node in this
case, which requires filtering out non-matching leaf nodes in the main
Search method since now searchHelper will return false positives for
prefix-matched leaves.